requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
mxnet>=1.8.0
scikit-learn>=0.24.0
scipy>=1.5.0
Pillow>=8.0.0
tqdm>=4.60.0
PyYAML>=6.0
